replace iframeLogin with windowOpenLogin
rationale: users are advised never to enter their password if the url in the browser is not the well known CAS url.

solution: open a new browser window instead of an iframe.

nb: browsers often block window.open, so display a message to the user then open the window just after user's click